feat: pricing calculator advances
